Fixed tests after introduction of IUnitySceneManagerWrapper as new input parameter for State
