+++ Fightship Arena/Assets/Scripts/Managers/GameManagement/StateMachine/IStack.cs
@@ -1,0 +1,13 @@
+﻿namespace FightShipArena.Assets.Scripts.Managers.GameManagement.StateMachine
+{
+    public interface IStack<T>
+    {
+        int Count { get; }
+        void Clear();
+        bool Contains(T item);
+        T Peek();
+        T Pop();
+        void Push(T item);
+        T[] ToArray();
+    }
+}
