+++ Fightship Arena/Assets/Scripts/Managers/GameManagement/StateMachine/IStack.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f3b87cf34bca30242a896a2108db8f37
+guid: 551fc8630378e2a40823fb3df1e93994
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
