+++ Fightship Arena/Assets/Scripts/Managers/GameManagement/StateMachine/TestableStack.cs
@@ -2,6 +2,12 @@
 
 namespace FightShipArena.Assets.Scripts.Managers.GameManagement.StateMachine
 {
+    /// <summary>
+    /// Bridge pattern is applied here to decouple the code from the actual implementation of the Stack class.
+    /// This is to improve testability in dependent classes.
+    /// This class, on the contrary, is not testable as the Stack class is not mockable.
+    /// </summary>
+    /// <typeparam name="T"></typeparam>
     public class TestableStack<T> : IStack<T>
     {
         protected Stack<T> _stack = new Stack<T>();
