--- Fightship Arena/Assets/Tests/EditMode/Managers/GameManagement/StateMachine/TestableStackUnitTests.cs
@@ -1,15 +1,0 @@
-﻿using NUnit.Framework;
-using NUnit.Framework.Internal;
-
-namespace FightshipArena.Assets.Tests.EditMode.Managers.GameManagement.StateMachine
-{
-    [TestFixture]
-    public class TestableStackUnitTests
-    {
-        [Test]
-        public void TestsMissing()
-        {
-            Assert.Fail("Tests missing");
-        }
-    }
-}
